+++ examples/svelte-kit/src/routes/index.svelte
@@ -1,12 +1,5 @@
-<script lang="ts">
-  import App from '../../../svelte/src/App.svelte'
+<script>
+  import App from '@bytemd-examples/svelte-webpack/src/App.svelte'
 </script>
 
 <App />
-
-<style lang="scss">
-  :root {
-    font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen,
-      Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;
-  }
-</style>