+++ packages/bytemd/src/editor.svelte
@@ -1,7 +1,7 @@
 <svelte:options immutable={true} />
 
 <script lang="ts">
-  import type { Editor, KeyMap, Linter, Annotation } from 'codemirror'
+  import type { Editor, KeyMap } from 'codemirror'
   import type { Root, Element } from 'hast'
   import type { VFile } from 'vfile'
   import type { BytemdEditorContext, BytemdPlugin, EditorProps } from './types'
@@ -29,9 +29,6 @@
   import useGfm from 'codemirror-ssr/mode/gfm/gfm'
   import useYaml from 'codemirror-ssr/mode/yaml/yaml'
   import useYamlFrontmatter from 'codemirror-ssr/mode/yaml-frontmatter/yaml-frontmatter'
-  import useVim from 'codemirror-ssr/keymap/vim'
-  import useEmacs from 'codemirror-ssr/keymap/emacs'
-  import useLint from 'codemirror-ssr/addon/lint/lint'
 
   export let value: EditorProps['value'] = ''
   export let plugins: NonNullable<EditorProps['plugins']> = []
@@ -43,9 +40,9 @@
   export let locale: EditorProps['locale']
   export let uploadImages: EditorProps['uploadImages']
   export let overridePreview: EditorProps['overridePreview']
-  export let maxLength: EditorProps['maxLength']
+  export let maxLength: NonNullable<EditorProps['maxLength']> = Infinity
 
-  const mergedLocale = { ...en, ...locale }
+  $: mergedLocale = { ...en, ...locale }
   const dispatch = createEventDispatcher()
 
   $: actions = getBuiltinActions(mergedLocale, plugins, uploadImages)
@@ -65,7 +62,6 @@
   let activeTab: false | 'write' | 'preview'
   let fullscreen = false
   let sidebar: false | 'help' | 'toc' = false
-  let islimited = false
 
   $: styles = (() => {
     let edit: string
@@ -126,7 +122,7 @@
     // console.log('off', plugins);
     cbs.forEach((cb) => cb && cb())
 
-    editor.removeKeyMap(keyMap)
+    editor?.removeKeyMap(keyMap) // onDestroy runs at SSR, optional chaining here
   }
 
   let debouncedValue = value
@@ -167,34 +163,6 @@
     useGfm(codemirror)
     useYaml(codemirror)
     useYamlFrontmatter(codemirror)
-    useVim(codemirror)
-    useEmacs(codemirror)
-    useLint(codemirror)
-
-    // lint
-    const linter: Linter = () => {
-      if (!vfile) return []
-
-      const annotations = vfile.messages.map((m) => {
-        const a: Annotation = {
-          from: codemirror.Pos(
-            m.location.start.line - 1,
-            m.location.start.column - 1
-          ),
-          to:
-            m.location.end.line == null // TODO: why null?
-              ? codemirror.Pos(m.location.start.line - 1)
-              : codemirror.Pos(
-                  m.location.end.line - 1,
-                  m.location.end.column - 1
-                ),
-          message: m.message,
-        }
-        return a
-      })
-      // console.log(annotations);
-      return annotations
-    }
 
     // @ts-ignore TODO: type
     editor = codemirror(editorEl, {
@@ -205,7 +173,6 @@
       indentUnit: 4, // nested ordered list does not work with 2 spaces
       ...editorConfig,
       placeholder,
-      lint: linter,
     })
 
     // https://github.com/codemirror/CodeMirror/issues/2428#issuecomment-39315423
@@ -215,20 +182,6 @@
       'Shift-Tab': 'indentLess',
     })
 
-    editor.on('beforeChange', (editor, change) => {
-      if (maxLength && change.update) {
-        let str = change.text.join('\n')
-        const to = editor.indexFromPos(change.to)
-        const from = editor.indexFromPos(change.from)
-        const offset =
-          editor.getValue().length + (str.length - (to - from)) - maxLength
-        islimited = offset >= 0
-        if (islimited) {
-          str = str.substr(0, str.length - offset)
-          change.update(change.from, change.to, str.split('\n'))
-        }
-      }
-    })
     editor.on('change', () => {
       dispatch('change', { value: editor.getValue() })
     })
@@ -242,8 +195,8 @@
       if (!(body instanceof HTMLElement)) return
 
       const leftNodes = hast.children.filter(
-        (v) => v.type === 'element'
-      ) as Element[]
+        (v): v is Element => v.type === 'element'
+      )
       const rightNodes = [...body.childNodes].filter(
         (v): v is HTMLElement => v instanceof HTMLElement
       )
@@ -471,7 +424,7 @@
     showSync={!overridePreview && split}
     value={debouncedValue}
     {syncEnabled}
-    {islimited}
+    islimited={value.length > maxLength}
     on:sync={(e) => {
       syncEnabled = e.detail
     }}