+++ packages/bytemd/src/status.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
   import type { BytemdLocale } from './types'
   import { createEventDispatcher } from 'svelte'
-  const wordCount = require('word-count')
+  import wordCount from 'word-count'
 
   export let showSync: boolean
   export let value: string
@@ -24,7 +24,7 @@
       {locale.lines}: <strong>{lines}</strong>
     </span>
     {#if islimited}
-      <span class="danger">{locale.limited}</span>
+      <span class="bytemd-status-error">{locale.limited}</span>
     {/if}
   </div>